# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Royal Dental Centre API
 * OpenAPI spec version: 0.1.0
 */
import type { PrescriptionItem } from "./prescriptionItem";

export interface Prescription {
  id: number;
  patientId: number;
  doctorId: number;
  patientName: string;
  /** @nullable */
  patientAge?: number | null;
  doctorName: string;
  items: PrescriptionItem[];
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}