# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Royal Dental Centre API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  nameAr: string;
  description: string;
  descriptionAr: string;
  duration: number;
  price: number;
  isActive: boolean;
  createdAt: Date;
}